added some evaluation experimental featrue

+++ src/BestMoveSearch.cpp
@@ -21,7 +21,7 @@ void BestMoveSearch::_insertionSort(std::vector<std::pair<int, int>>& list)
 
 void BestMoveSearch::_heapSortMoves(MoveGenerator::payload moves)
 {
-    for (ssize_t i = 1; i < moves.size; ++i)
+    for (ssize_t i = 1; i < static_cast<ssize_t>(moves.size); ++i)
     {
         ssize_t j = i - 1;
         const auto val = moves[i];
